requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
blas=1.0=mkl
bottleneck=1.3.5=py311hbed6279_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.05.30=h06a4308_0
click=8.0.4=py311h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46305
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
mkl=2023.1.0=h6d00ec8_46342
mkl-service=2.4.0=py311h5eee18b_1
mkl_fft=1.3.6=py311ha02d727_1
mkl_random=1.2.2=py311ha02d727_1
ncurses=6.4=h6a678d5_0
numexpr=2.8.4=py311h65dcdc2_1
numpy=1.25.0=py311h08b1b3b_0
numpy-base=1.25.0=py311hf175353_0
openssl=3.0.9=h7f8727e_0
pandas=1.5.3=py311hba01205_0
pip=23.1.2=py311h06a4308_0
pymongo=3.12.0=py311h6a678d5_0
python=3.11.4=h955ad1f_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-dotenv=0.21.0=py311h06a4308_0
pytz=2022.7=py311h06a4308_0
readline=8.2=h5eee18b_0
setuptools=67.8.0=py311h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h5eee18b_0
tbb=2021.8.0=hdb19cb5_0
tk=8.6.12=h1ccaba5_0
tzdata=2023c=h04d1e81_0
wheel=0.38.4=py311h06a4308_0
xz=5.4.2=h5eee18b_0
zlib=1.2.13=h5eee18b_0
